requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
